pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ravipangali-sites-status-attacker"
version = "0.1.0"
description = "A django package made by Ravi Pangali for down those sites which are made by himself"
readme = "README.md"
requires-python = ">=3.12.4"
license = {file = "LICENSE"}
authors = [
    {name = "Ravi Pangali", email = "legendromeoravi@gmail.com"},
]
dependencies = [
    "Django>=5.2.1",
    "djangorestframework>=3.16.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: Django",
]

[project.urls]
"Homepage" = "https://github.com/ravipangali/RaviPangali-Site-Status-Attacker"
"Repository" = "https://github.com/ravipangali/RaviPangali-Site-Status-Attacker"